touch ups

+++ contracts/BIFI/interfaces/common/IERC20Extended.sol
@@ -4,5 +4,6 @@ pragma solidity >=0.6.0 <0.9.0;
 
 interface IERC20Extended {
     function symbol() external view returns (string memory);
-    function decimals() external view returns (uint);
+
+    function decimals() external view returns (uint256);
 }
